# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Viveris Carbone API",
    "description": "FastAPI backend for Viveris Carbone project",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a new user account",
        "description": "Create a new user with `email`, `user_name` and `password`. Returns the created user (no password).",
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login and receive a JWT",
        "description": "Authenticate with `email` and `password`. Returns an access token on success.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/emissions/save": {
      "post": {
        "tags": [
          "emissions"
        ],
        "summary": "Save user footprint predictions",
        "description": "Calculates the true footprint based on submitted answers and saves it.",
        "operationId": "save_emissions_api_emissions_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmissionsSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmissionsSaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/emissions/category": {
      "put": {
        "tags": [
          "emissions"
        ],
        "summary": "Recalculate a specific category footprint",
        "description": "Calculates a new total for a specific category and maintains history.",
        "operationId": "update_category_emissions_api_emissions_category_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/users/me/friends": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "Get user's friends",
        "description": "Returns the list of friends for the currently authenticated user.",
        "operationId": "get_friends_api_users_me_friends_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendResponse"
                  },
                  "type": "array",
                  "title": "Response Get Friends Api Users Me Friends Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/{friend_id}/friends": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Send a friend request",
        "description": "Sends a friend request to another user. If they already requested you, it auto-accepts.",
        "operationId": "send_friend_request_api_users__friend_id__friends_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Friend Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_friend_request_api_users__friend_id__friends_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "friends"
        ],
        "summary": "Remove a friend or cancel request",
        "operationId": "delete_friend_api_users__friend_id__friends_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Friend Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{friend_id}/friends/accept": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Accept a friend request",
        "operationId": "accept_friend_request_api_users__friend_id__friends_accept_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Friend Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Requestid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{friend_id}/friends/decline": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Decline a friend request",
        "operationId": "decline_friend_request_api_users__friend_id__friends_decline_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Friend Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Reason"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerItem": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "co2": {
            "type": "number",
            "title": "Co2"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value",
          "co2"
        ],
        "title": "AnswerItem"
      },
      "Body_send_friend_request_api_users__friend_id__friends_post": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "Body_send_friend_request_api_users__friend_id__friends_post"
      },
      "CategoryUpdateRequest": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "answers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AnswerItem"
            },
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "category",
          "answers"
        ],
        "title": "CategoryUpdateRequest"
      },
      "CategoryUpdateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "newCategoryTotal": {
            "type": "number",
            "title": "Newcategorytotal"
          },
          "categoryEmissions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Categoryemissions"
          },
          "updatedTotalInTons": {
            "type": "number",
            "title": "Updatedtotalintons"
          },
          "savedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Savedat"
          }
        },
        "type": "object",
        "required": [
          "category",
          "newCategoryTotal",
          "categoryEmissions",
          "updatedTotalInTons",
          "savedAt"
        ],
        "title": "CategoryUpdateResponse"
      },
...
      "EmissionsSaveRequest": {
        "properties": {
          "quizResult": {
            "$ref": "#/components/schemas/QuizResultPayload"
          },
          "userPredictions": {
            "$ref": "#/components/schemas/UserPredictions"
          }
        },
        "type": "object",
        "required": [
          "quizResult",
          "userPredictions"
        ],
        "title": "EmissionsSaveRequest"
      },
      "EmissionsSaveResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "emissionId": {
            "type": "string",
            "title": "Emissionid"
          },
          "quizResult": {
            "$ref": "#/components/schemas/QuizResultResponse"
          },
          "categoryEmissions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Categoryemissions"
          },
          "targetCO2": {
            "type": "number",
            "title": "Targetco2"
          },
          "savedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Savedat"
          }
        },
        "type": "object",
        "required": [
          "emissionId",
          "quizResult",
          "categoryEmissions",
          "targetCO2",
          "savedAt"
        ],
        "title": "EmissionsSaveResponse"
      },
      "FriendActionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "friendId": {
            "type": "string",
            "title": "Friendid"
          },
          "totalFriends": {
            "type": "integer",
            "title": "Totalfriends",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "status",
          "friendId"
        ],
        "title": "FriendActionResponse"
      },
      "FriendResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "trees": {
            "type": "integer",
            "title": "Trees"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "avatar",
          "points",
          "trees",
          "status"
        ],
        "title": "FriendResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "example": {
          "email": "alice@example.com",
          "password": "strong_password_123"
        }
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "LoginResponse",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "token_type": "bearer"
        }
      },
...
      "QuizResultPayload": {
        "properties": {
          "totalInTons": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "title": "Totalintons"
          },
          "answers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AnswerItem"
            },
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "totalInTons",
          "answers"
        ],
        "title": "QuizResultPayload"
      },
      "QuizResultResponse": {
        "properties": {
          "totalInTons": {
            "type": "number",
            "title": "Totalintons"
          },
          "categoryBreakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Categorybreakdown"
          }
        },
        "type": "object",
        "required": [
          "totalInTons",
          "categoryBreakdown"
        ],
        "title": "QuizResultResponse"
      },
      "SignUpRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Password (minimum 8 characters)"
          },
          "user_name": {
            "type": "string",
            "minLength": 3,
            "title": "User Name",
            "description": "Unique username"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "user_name"
        ],
        "title": "SignUpRequest",
        "example": {
          "email": "alice@example.com",
          "password": "strong_password_123",
          "user_name": "alice"
        }
      },
      "SignUpResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "user_name"
        ],
        "title": "SignUpResponse",
        "example": {
          "email": "alice@example.com",
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "user_name": "alice"
        }
      },
...
      "UserPredictions": {
        "properties": {
          "highestConsumption": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highestconsumption"
          },
          "flexibility": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flexibility"
          }
        },
        "type": "object",
        "required": [
          "flexibility"
        ],
        "title": "UserPredictions"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}